+++ qos/solana/response_gethealth.go
@@ -69,7 +69,7 @@ type responseToGetHealth struct {
 func (r responseToGetHealth) GetObservation() qosobservations.SolanaEndpointObservation {
 	return qosobservations.SolanaEndpointObservation{
 		// Set the HTTP status code using the JSONRPC Response
-		HttpStatusCode: r.Response.GetRecommendedHTTPStatusCode(),
+		HttpStatusCode: int32(r.Response.GetRecommendedHTTPStatusCode()),
 		ResponseObservation: &qosobservations.SolanaEndpointObservation_GetHealthResponse{
 			GetHealthResponse: &qosobservations.SolanaGetHealthResponse{
 				Result: r.HealthResult,
